requirements.txt
openai==1.40.2
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
pinecone-client==5.0.1
pydantic==2.8.2
